Client rework with home done.
Leaving the electron aside for a bit, the pure npm website now has a working home page. The home page allows you to either load the example, or drag and drop / browse a file from your computer.

The tabs and pages now have a way to be open and closed.

This introduced the Grommet UI library too.